+++ projects/igniteui-angular/src/lib/grids/filtering/excel-style/excel-style-conditional-filter.component.ts
@@ -165,6 +165,7 @@ export class IgxExcelStyleConditionalFilterComponent implements OnDestroy {
             case DataType.Boolean:
                 return this.esf.grid.resourceStrings.igx_grid_excel_boolean_filter;
             case DataType.Number:
+            case DataType.Percent:
                 return this.esf.grid.resourceStrings.igx_grid_excel_number_filter;
             case DataType.Date:
                 return this.esf.grid.resourceStrings.igx_grid_excel_date_filter;