+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes for each version of this project will be documented in this 
         ```
 - `IgxGrid`, `IgxTreeGrid`, `IgxHierarchicalGrid`
     - Support for `currency` type columns is added in the grid.
+    - Support for `percent` type columns is added in the grid.
     - Added support for filtering based on the formatted cell values using the `FormattedValuesFilteringStrategy` for `IgxGrid`/`IgxHierarchicalGrid` and `TreeGridFormattedValuesFilteringStrategy` for `IgxTreeGrid`.
 - `IgxPaginator`
     - `paging` and `pagingDone` events are now emitted.
@@ -38,6 +39,7 @@ All notable changes for each version of this project will be documented in this 
 - `IgxDialog`
     - The dialog content has been moved inside the dialog window container in the template. This means that if you have added something in-between the opening and closing tags of the dialog, you may have to adjust its styling a bit since that content is now rendered inside a container that has padding on it.
 - `IgxCalendar`
+    - **Breaking Change**
     - A new string enumeration `IgxCalendarView` is exported. Either the new one or the current `CalendarView` can be used. `CalendarView` will be deprecated in a future release.
     - `onSelection` is now `selected`
     - `onViewChanging` is now `viewChanging`
@@ -45,15 +47,19 @@ All notable changes for each version of this project will be documented in this 
     - `onYearSelection` is now `yearSelection`
     - `onMonthSelection` is now `monthSelection`
 - `IgxYearsViewComponent`
+    - **Breaking Change**
     - `onSelection` is now `selected`
     - `onYearSelection` is now `yearSelection`
 - `IgxDaysViewComponent`
+    - **Breaking Change**
     - `onDateSelection` is now `dateSelection`
     - `onViewChanging` is now `viewChanging`
 - `IgxMonthsViewComponent`
+    - **Breaking Change**
     - `onSelection` is now `selected`
     - `onMonthSelection` is now `monthSelection`
 - `IgxMonthPickerComponent`
+    - **Breaking Change**
     - `onSelection` is now `selected`
 - `IgxRadioGroup`
     - Added new property `alignment` that determines the radio group alignment. Available options are `horizontal` (default) and `vertical`.